Chore: rename implementation class name
