+++ gradle-plugin/src/main/java/dev/rikka/tools/refine/RefinePlugin.java
@@ -1,4 +1,4 @@
-package dev.rikka.tools;
+package dev.rikka.tools.refine;
 
 import com.android.build.gradle.BaseExtension;
 import org.gradle.api.GradleException;
@@ -6,13 +6,13 @@ import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 
 @SuppressWarnings("unused")
-public class HiddenApiRefinePlugin implements Plugin<Project> {
+public class RefinePlugin implements Plugin<Project> {
     @Override
     public void apply(Project target) {
         BaseExtension androidExtension = target.getExtensions().findByType(BaseExtension.class);
         if (androidExtension == null)
             throw new GradleException("Android extension not found");
 
-        androidExtension.registerTransform(new HiddenApiRefineTransform());
+        androidExtension.registerTransform(new RefineTransform());
     }
 }
